feat: improve distance calc func code

+++ domain/position.go
@@ -4,5 +4,5 @@ type Position struct {
 	Name      string  `json:"name,omitempty" valid:"-"`
 	Latitude  float64 `json:"latitude,omitempty" valid:"latitude"`
 	Longitude float64 `json:"longitude,omitempty" valid:"longitude"`
-	Distance  float64 `json:"distance,omitempty" valid:"-"`
+	Distance  uint16  `json:"distance,omitempty" valid:"-"`
 }
